requirements.txt
pyserial
minimalmodbus
pymodbus
numpy